+++ src/SFA.DAS.LevyTransferMatching.Web/Extensions/DoubleExtensions.cs
@@ -4,9 +4,9 @@ namespace SFA.DAS.LevyTransferMatching.Web.Extensions
 {
     public static class DoubleExtensions
     {
-        public static double ToNearest(this double value, int roundTo)
+        public static int ToNearest(this double value, int roundTo)
         {
-            return Math.Round(value / roundTo, MidpointRounding.AwayFromZero) * roundTo;
+            return (int)Math.Round(value / roundTo, MidpointRounding.AwayFromZero) * roundTo;
         }
     }
 }
